+++ hypella_indicators/__init__.py
@@ -1,4 +1,4 @@
-from .core import Candle, Indicator
+from .core import CandleData, Indicator
 from .indicators.rsi import RSI
 from .indicators.rsi_sma import RSISMA
 from .indicators.sma import SMA
@@ -7,5 +7,10 @@ from .indicators.volume_sma import VolumeSMA
 from .indicators.bb import BollingerBands
 from .indicators.atr import ATR
 from .indicators.adx import ADX
+from .indicators.candle import Candle
+from .indicators.price import Price
 
-__all__ = ["Candle", "Indicator", "RSI", "RSISMA", "SMA", "EMA", "VolumeSMA", "BollingerBands", "ATR", "ADX"]
+__all__ = [
+    "CandleData", "Indicator", "RSI", "RSISMA", "SMA", "EMA", 
+    "VolumeSMA", "BollingerBands", "ATR", "ADX", "Candle", "Price"
+]